@@ -5,7 +5,7 @@ version: 0.1.1
 homepage: https://github.com/jhourlad/flutter_session
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
